다형성을 이용해 중복 분기 제거
